chore: Adjusted Dockerfile and configure.js to use external directories if necessary (#237)

+++ bin/configure.js
@@ -32,21 +32,25 @@ const [defaultGitConfigurationUrl, defaultGitConfigurationVersion] = pkg.depende
 );
 const gitConfigurationUrl = process.env.WIRE_CONFIGURATION_REPOSITORY || defaultGitConfigurationUrl;
 const gitConfigurationVersion = process.env.WIRE_CONFIGURATION_REPOSITORY_VERSION || defaultGitConfigurationVersion;
-
-console.log(
-  `Loading configuration version "${gitConfigurationVersion}" for project "${pkg.name}" from "${gitConfigurationUrl}"`
-);
-const configDirName = 'config';
+const configDirName = process.env.WIRE_CONFIGURATION_EXTERNAL_DIR || 'config';
 const configDir = resolve(configDirName);
 const src = resolve(configDir, pkg.name, 'content');
 const root = '.';
 const dest = `../../../dist`;
 const ignoreList = ['.DS_Store'];
 
-
-console.log(`Cleaning config directory "${configDir}"`);
-fs.removeSync(configDir);
-execSync(`git clone --single-branch -b ${gitConfigurationVersion} ${gitConfigurationUrl} ${configDirName}`, {stdio: [0, 1]});
+if (!process.env.WIRE_CONFIGURATION_EXTERNAL_DIR) {
+  console.log(
+    `Loading configuration version "${gitConfigurationVersion}" for project "${
+      pkg.name
+    }" from "${gitConfigurationUrl}" \
+     and cleaning config directory "${configDir}"`
+  );
+  fs.removeSync(configDir);
+  execSync(`git clone -b ${gitConfigurationVersion} ${gitConfigurationUrl} ${configDirName}`, {stdio: [0, 1]});
+} else {
+  console.log(`Using external config directory "${configDir}"`);
+}
 
 // Copy .env file configuration
 console.log('env', resolve(configDir, pkg.name, '.env'), resolve(root, '.env'));
